+++ src/Plane/Polygons/AbstractPolygon.php
@@ -2,13 +2,14 @@
 namespace samizdam\Geometry\Plane\Polygons;
 
 use samizdam\Geometry\Plane\CalculatorAwareTrait;
+use samizdam\Geometry\Plane\CompositeCalculatorAwareInterface;
 
 /**
  *
  * @author samizdam
  *        
  */
-abstract class AbstractPolygon implements PolygonInterface
+abstract class AbstractPolygon implements PolygonInterface, CompositeCalculatorAwareInterface
 {
     
     use CalculatorAwareTrait;
@@ -28,6 +29,20 @@ abstract class AbstractPolygon implements PolygonInterface
             $polygon = new Polygon($points);
         }
         return $polygon;
+    }
+
+    public function getArea()
+    {
+        return $this->getCompositeCalculator()
+            ->getCalculator(PolygonCalculatorInterface::class)
+            ->getArea($this);
+    }
+
+    public function getLength()
+    {
+        return $this->getCompositeCalculator()
+            ->getCalculator(PolygonCalculatorInterface::class)
+            ->getLength($this);
     }
 
     public function getPoints()