+++ src/Plane/Polygons/PolygonCalculatorInterface.php
@@ -13,6 +13,14 @@ interface PolygonCalculatorInterface
      * Get area of Polygon
      *
      * @param PolygonInterface $polygon
+     * @return float
      */
     public function getArea(PolygonInterface $polygon);
+    
+    /**
+     * 
+     * @param PolygonInterface $polygon
+     * @return float
+     */
+    public function getLength(PolygonInterface $polygon);
 }