+++ src/Plane/Polygons/PolygonInterface.php
@@ -1,10 +1,31 @@
 <?php
 namespace samizdam\Geometry\Plane\Polygons;
 
+use samizdam\Geometry\Plane\PointInterface;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
 interface PolygonInterface
 {
 
-    public function getPoints();
-
+    /**
+     *
+     * @return float
+     */
     public function getArea();
+
+    /**
+     *
+     * @return float
+     */
+    public function getLength();
+
+    /**
+     *
+     * @return PointInterface[]
+     */
+    public function getPoints();
 }