+++ src/Plane/Polygons/Triangle.php
@@ -3,12 +3,11 @@ namespace samizdam\Geometry\Plane\Polygons;
 
 use samizdam\Geometry\Plane\CalculatorAwareTrait;
 
+/**
+ * 
+ * @author samizdam
+ *
+ */
 class Triangle extends AbstractPolygon
 {
-    use CalculatorAwareTrait;
-
-    public function getArea()
-    {
-        return $this->getCompositeCalculator()->getCalculator(PolygonCalculatorInterface::class)->getArea($this);
-    }
 }