+++ tests/Plane/Polygons/PolygonTest.php
@@ -52,4 +52,15 @@ class PolygonTest extends GeometryUnitTestCase
         $polygon = AbstractPolygon::createByPoints($points);
         $this->assertEquals($points, $polygon->getPoints());
     }
+    
+    public function testGetLength()
+    {
+        $polygon = AbstractPolygon::createByPoints([
+            new Point(0, 0),
+            new Point(0, 1),
+            new Point(1, 1),
+            new Point(1, 0),
+        ]);
+        $this->assertEquals(4, $polygon->getLength());
+    }
 }