+++ CHANGELOG.md
@@ -18,7 +18,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Normalize naming of arguments, fields and methods. To fix in some guideline.
- 
+
+## 0.5.2 - 2015-09-XX 
+### Added
+- Polygon::getLength() for perimeter. 
+- CollectionInterface over abstract class.  
+- ::toArray() for collections. 
+
+### Fixed
+- Missing interface in AbstractPolygon for correct inject Calc. 
+- Remove dublicated code from Polygon implementations. 
 
 ## 0.5.1 - 2015-09-23
 ### Changed